Create new-user.ps1
This PowerShell script allows you to create a new user account in Active Directory and add the user to a specified group. The script prompts you to enter the user's first name, last name, username, user description, and group name. It then generates a random password for the user and displays it in the output message.